+++ libs/core/src/lib/token/token.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { TokenComponent } from './token.component';
+import { TokenizerComponent } from './tokenizer.component';
+import { TokenizerInputDirective } from './token-input.directive';
 
 @NgModule({
-    declarations: [TokenComponent],
+    declarations: [TokenComponent, TokenizerComponent, TokenizerInputDirective],
     imports: [CommonModule],
-    exports: [TokenComponent]
+    exports: [TokenComponent, TokenizerComponent, TokenizerInputDirective]
 })
 export class TokenModule {}