+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.14.0-rc.36"></a>
+# [0.14.0-rc.36](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.35...v0.14.0-rc.36) (2020-02-11)
+
+
+### Bug Fixes
+
+* token and tokenizer implementation of 0.5.0 ([#1878](https://github.com/SAP/fundamental-ngx/issues/1878)) ([dd2e8ad](https://github.com/SAP/fundamental-ngx/commit/dd2e8ad))
+
+
+
+<a name="0.14.0-rc.35"></a>
+# [0.14.0-rc.35](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.34...v0.14.0-rc.35) (2020-02-11)
+
+
+### Features
+
+* (platform) Link as a standalone component ([#1903](https://github.com/SAP/fundamental-ngx/issues/1903)) ([dd68eb9](https://github.com/SAP/fundamental-ngx/commit/dd68eb9))
+
+
+
+<a name="0.14.0-rc.34"></a>
+# [0.14.0-rc.34](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.33...v0.14.0-rc.34) (2020-02-10)
+
+
+
+<a name="0.14.0-rc.33"></a>
+# [0.14.0-rc.33](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.32...v0.14.0-rc.33) (2020-02-10)
+
+
+
+<a name="0.14.0-rc.32"></a>
+# [0.14.0-rc.32](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.31...v0.14.0-rc.32) (2020-02-10)
+
+
+
 <a name="0.14.0-rc.31"></a>
 # [0.14.0-rc.31](https://github.com/SAP/fundamental-ngx/compare/v0.14.0-rc.30...v0.14.0-rc.31) (2020-02-06)